feat: move GMA compatibility check to static method
