+++ Example/PrebidDemoKotlin/src/main/java/org/prebid/mobile/prebidkotlindemo/MainActivity.kt
@@ -22,6 +22,8 @@ import android.widget.AdapterView
 import android.widget.ArrayAdapter
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
+import com.google.android.gms.ads.MobileAds
+import org.prebid.mobile.PrebidMobile
 import org.prebid.mobile.prebidkotlindemo.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
@@ -43,6 +45,8 @@ class MainActivity : AppCompatActivity() {
         binding.btnShowAd.setOnClickListener { showAd() }
         initAdServerSpinner()
         initDefaultServer()
+
+        PrebidMobile.checkGoogleMobileAdsCompatibility(MobileAds.getVersion().toString())
     }
 
     private fun showAd() {
